[ADD] #29 error management to recipe controller

+++ src/controllers/recipe.controller.js
@@ -43,6 +43,11 @@ class RecipeController {
           },
         ],
       });
+      if(!recipe){
+        const error = new Error('Bad request. Invalid ID');
+        error.status = 400;
+        throw error;
+      }
       res.status(200).json(recipe);
     } catch (err) {
       res
@@ -84,6 +89,14 @@ class RecipeController {
   update = async (req, res) => {
     const transaction = await db.sequelize.transaction();
     try {
+      const recipe = await db.Recipe.findByPk(req.params.id);
+
+      if(!recipe){
+        const error = new Error('Bad request. Invalid ID');
+        error.status = 400;
+        throw error;
+      }
+
       await db.Recipe.update(
         {
           title: req.body.title,
@@ -97,8 +110,6 @@ class RecipeController {
           transaction,
         }
       );
-
-      const recipe = await db.Recipe.findByPk(req.params.id);
 
       await transaction.commit();
       res
@@ -119,11 +130,20 @@ class RecipeController {
    */
   delete = async (req, res, next) => {
     try {
+      const recipe = await db.Recipe.findByPk(req.params.id);
+
+      if(!recipe){
+        const error = new Error('Bad request. Invalid ID');
+        error.status = 400;
+        throw error;
+      }
+
       const deleteRecipe = await db.Recipe.destroy({
         where: {
           id: req.params.id,
         },
       });
+
       res
         .status(200)
         .json({ message: `product with id ${req.params.id} deleted` });
